\name{ShapePattern-package}
\alias{ShapePattern-package}
\alias{ShapePattern}
\docType{package}
\title{
\packageTitle{ShapePattern}
}

\description{
\packageDescription{ShapePattern}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShapePattern}
\packageIndices{ShapePattern}
Perform ShrinkShape, a spatial decomposition by iterative shrinking of planar shapes (ploygons).
}

\author{
\packageAuthor{ShapePattern}

Maintainer: \packageMaintainer{ShapePattern}
}

\references{
Remmel, T.K. 2015. ShrinkShape2: a FOSS toolbox for computing rotation-invariant shape spectra for characterizing and comparing polygons. The Canadian Geographer 59(4):532-547.
}

\keyword{ package }
\keyword{ manip }

\examples{
# Something to be completed.
}
