\name{wi}
\alias{wi}
\title{
An internal utility that helps with the Whittle estimation.
}
\description{
This is an internal utility function for helping with the Whittle estimation. This function is not intended for use directly by users.
}
\usage{
wi(BE=data$demoimage1, CONTROL=FALSE, SIZE=6)
}
\arguments{
  \item{BE}{
This is a binary input image.
}
  \item{CONTROL}{
This is a Boolean flag.
}
  \item{SIZE}{
Controls the output matrix size.
}
}
\details{
This tool is only used internally.
}
\value{
The result is a value.
}
\references{
No references currently.
}
\author{
Ferko Csillag, Sandor Kabos, Tarmo K. Remmel
}
\note{
This function is not to be used directly by users of this package.
}
\seealso{
Currently none.
}
\examples{
# No example.
}
\keyword{misc}
